Refactor main window state handling
