+++ src/actions.js
@@ -50,8 +50,6 @@ export const SPELL_CHECKING_DICTIONARIES_UPDATED = 'spell-checking/dictionaries-
 export const SPELL_CHECKING_READY = 'spell-checking/ready';
 export const TOUCH_BAR_FORMAT_BUTTON_TOUCHED = 'touch-bar/format-button-touched';
 export const TOUCH_BAR_SELECT_SERVER_TOUCHED = 'touch-bar/select-server-touched';
-export const TRAY_ICON_CREATED = 'tray-icon/created';
-export const TRAY_ICON_DESTROYED = 'tray-icon/destroyed';
 export const TRAY_ICON_QUIT_CLICKED = 'tray-icon/quit-clicked';
 export const TRAY_ICON_TOGGLE_CLICKED = 'tray-icon/toggle-clicked';
 export const UPDATE_DIALOG_DISMISSED = 'update-dialog/dismissed';
