+++ src/reducers/openDialog.js
@@ -7,6 +7,9 @@ import {
 	SELECT_CLIENT_CERTIFICATE_DIALOG_CERTIFICATE_SELECTED,
 	SELECT_CLIENT_CERTIFICATE_DIALOG_DISMISSED,
 	UPDATE_DIALOG_DISMISSED,
+	UPDATE_DIALOG_DOWNLOAD_UPDATE_CLICKED,
+	UPDATE_DIALOG_REMIND_UPDATE_LATER_CLICKED,
+	UPDATE_DIALOG_SKIP_UPDATE_CLICKED,
 	UPDATES_NEW_VERSION_AVAILABLE,
 	WEBVIEW_SCREEN_SHARING_SOURCE_REQUESTED,
 } from '../actions';
@@ -31,6 +34,9 @@ export const openDialog = (state = null, { type }) => {
 		case SELECT_CLIENT_CERTIFICATE_DIALOG_CERTIFICATE_SELECTED:
 		case SELECT_CLIENT_CERTIFICATE_DIALOG_DISMISSED:
 		case UPDATE_DIALOG_DISMISSED:
+		case UPDATE_DIALOG_SKIP_UPDATE_CLICKED:
+		case UPDATE_DIALOG_REMIND_UPDATE_LATER_CLICKED:
+		case UPDATE_DIALOG_DOWNLOAD_UPDATE_CLICKED:
 			return null;
 	}
